HashMap은 Map의 종류이므로 HashMap을 Map 타입으로 변경 / Map Value값에 리스트가 아닌 Product 객체를 담는 방식으로 하여 진짜 DB 로직 구현
