--- src/main/java/com/garden/coffee/coffeegarden/repository/ProductRepository.java
@@ -1,12 +1,0 @@
-package com.garden.coffee.coffeegarden.repository;
-
-
-import com.garden.coffee.coffeegarden.entity.Product;
-
-import java.util.ArrayList;
-
-public interface ProductRepository {
-
-    abstract Product save(Product product);
-    abstract ArrayList<Product> search(Long productId);
-}
